Add guessed round dates in Krause export
This adds a method to guess sensible round dates for tournaments
if the start date, end date and number of rounds are set, and outputs
them when exporting to Krause format. These otherwise have to be added
manually before uploading to FIDE.

While ICU::Tournament has an attribute for round_dates, this gets
thrown away in the database layer, so we can't currently use the
dates from the uploaded file.